created folder helpers for posts function
